refactor(evidence): extract shared audit package

Copilot-Session: 13121eec-6224-47ef-8bae-baf1f9c4e090
