+++ .github/workflows/publish.yml
@@ -9,7 +9,7 @@ on:
         description: >
           Package to publish. Use "all" for everything, "all-python" for all
           Python packages, "all-npm" for all npm packages, or a specific name
-          (e.g. "agent-governance-toolkit-core", "agent-governance-copilot-cli", "agent-governance-claude-code", "agent-governance-opencode", "agent-governance-antigravity-cli", "agt-sandbox").
+          (e.g. "agt-evidence", "agent-governance-toolkit-core", "agent-governance-copilot-cli", "agent-governance-claude-code", "agent-governance-opencode", "agent-governance-antigravity-cli", "agt-sandbox").
         required: true
         type: string
         default: "all"
@@ -61,8 +61,11 @@ jobs:
     runs-on: ubuntu-latest
     outputs:
       matrix: ${{ steps.resolve.outputs.matrix }}
+      publish_matrix: ${{ steps.resolve.outputs.publish_matrix }}
       any: ${{ steps.resolve.outputs.any }}
       build_any: ${{ steps.resolve.outputs.build_any }}
+      publish_any: ${{ steps.resolve.outputs.publish_any }}
+      evidence: ${{ steps.resolve.outputs.evidence }}
       acs: ${{ steps.resolve.outputs.acs }}
     steps:
       - name: Resolve Python package matrix
@@ -73,6 +76,7 @@ jobs:
         run: |
           # Complete list of publishable Python packages.
           ALL='[
+            {"name":"agt-evidence","path":"agent-governance-python/agt-evidence"},
             {"name":"agent-governance-toolkit-core","path":"agent-governance-python/agent-governance-toolkit-core"},
             {"name":"agent-governance-toolkit-integrations","path":"agent-governance-python/agent-governance-toolkit-integrations"},
             {"name":"agent-governance-toolkit-cli","path":"agent-governance-python/agent-governance-toolkit-cli"},
@@ -133,19 +137,38 @@ jobs:
               fi
             else
               SELECTED="$FILTERED"
+              EVIDENCE_DEPENDENTS='[
+                "agent-governance-toolkit-core",
+                "agent-governance-toolkit-integrations",
+                "agent-governance-toolkit-cli",
+                "agent-governance-toolkit-protocols",
+                "agent-governance-toolkit"
+              ]'
+              if echo "$EVIDENCE_DEPENDENTS" | jq -e --arg name "$INPUT" 'index($name)' >/dev/null; then
+                EVIDENCE=$(echo "$ALL" | jq -c '[.[] | select(.name == "agt-evidence")]')
+                SELECTED=$(echo "$SELECTED" | jq -c --argjson evidence "$EVIDENCE" '$evidence + .')
+              fi
             fi
           fi
 
           GENERIC=$(echo "$SELECTED" | jq -c '[.[] | select(.name != "agent-control-specification")]')
+          PUBLISH=$(echo "$GENERIC" | jq -c '[.[] | select(.name != "agt-evidence")]')
           SELECTED_COUNT=$(echo "$SELECTED" | jq 'length')
           GENERIC_COUNT=$(echo "$GENERIC" | jq 'length')
+          PUBLISH_COUNT=$(echo "$PUBLISH" | jq 'length')
+          EVIDENCE_COUNT=$(echo "$GENERIC" | jq '[.[] | select(.name == "agt-evidence")] | length')
           ACS_COUNT=$(echo "$SELECTED" | jq '[.[] | select(.name == "agent-control-specification")] | length')
           any=$([ "$SELECTED_COUNT" -gt 0 ] && echo true || echo false)
           build_any=$([ "$GENERIC_COUNT" -gt 0 ] && echo true || echo false)
+          publish_any=$([ "$PUBLISH_COUNT" -gt 0 ] && echo true || echo false)
+          evidence=$([ "$EVIDENCE_COUNT" -eq 1 ] && echo true || echo false)
           acs=$([ "$ACS_COUNT" -eq 1 ] && echo true || echo false)
           echo "matrix={\"include\":$GENERIC}" >> "$GITHUB_OUTPUT"
+          echo "publish_matrix={\"include\":$PUBLISH}" >> "$GITHUB_OUTPUT"
           echo "any=$any" >> "$GITHUB_OUTPUT"
           echo "build_any=$build_any" >> "$GITHUB_OUTPUT"
+          echo "publish_any=$publish_any" >> "$GITHUB_OUTPUT"
+          echo "evidence=$evidence" >> "$GITHUB_OUTPUT"
           echo "acs=$acs" >> "$GITHUB_OUTPUT"
 
   acs-python-wheels:
@@ -408,9 +431,70 @@ jobs:
         with:
           subject-path: ${{ matrix.path }}/dist/*
 
-      - name: Prepare PyPI upload artifacts
-        if: github.event_name == 'workflow_dispatch' && github.event.inputs.dry_run == 'false'
-        working-directory: ${{ matrix.path }}
+      - name: Upload build artifacts
+        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
+        with:
+          name: pypi-${{ matrix.name }}
+          path: ${{ matrix.path }}/dist/
+          retention-days: 30
+
+  publish-agt-evidence:
+    needs: [resolve-python-matrix, build-python]
+    if: >-
+      github.event_name == 'workflow_dispatch' &&
+      github.event.inputs.dry_run == 'false' &&
+      needs.resolve-python-matrix.outputs.evidence == 'true'
+    runs-on: ubuntu-latest
+    environment: pypi
+    permissions:
+      contents: read
+      id-token: write
+    steps:
+      - uses: actions/download-artifact@3e5f45b2cfb9172054b4087a40e8e0b5a5461e7c # v8.0.1
+        with:
+          name: pypi-agt-evidence
+          path: dist
+      - name: Prepare agt-evidence upload artifacts
+        run: |
+          rm -rf pypi-upload
+          mkdir -p pypi-upload
+          shopt -s nullglob
+          artifacts=(dist/*.whl dist/*.tar.gz)
+          if [ "${#artifacts[@]}" -eq 0 ]; then
+            echo "::error::No wheel or sdist artifacts found for agt-evidence"
+            exit 1
+          fi
+          cp "${artifacts[@]}" pypi-upload/
+      - name: Publish to PyPI
+        uses: pypa/gh-action-pypi-publish@dc37677b2e1c63e2034f94d8a5b11f265b73ba33 # v1.14.2
+        with:
+          packages-dir: pypi-upload/
+          skip-existing: true
+
+  publish-python:
+    needs: [resolve-python-matrix, build-python, publish-agt-evidence]
+    if: >-
+      always() &&
+      github.event_name == 'workflow_dispatch' &&
+      github.event.inputs.dry_run == 'false' &&
+      needs.build-python.result == 'success' &&
+      needs.resolve-python-matrix.outputs.publish_any == 'true' &&
+      (needs.resolve-python-matrix.outputs.evidence != 'true' ||
+       needs.publish-agt-evidence.result == 'success')
+    runs-on: ubuntu-latest
+    environment: pypi
+    permissions:
+      contents: read
+      id-token: write
+    strategy:
+      fail-fast: false
+      matrix: ${{ fromJson(needs.resolve-python-matrix.outputs.publish_matrix) }}
+    steps:
+      - uses: actions/download-artifact@3e5f45b2cfb9172054b4087a40e8e0b5a5461e7c # v8.0.1
+        with:
+          name: pypi-${{ matrix.name }}
+          path: dist
+      - name: Prepare ${{ matrix.name }} upload artifacts
         run: |
           rm -rf pypi-upload
           mkdir -p pypi-upload
@@ -421,19 +505,10 @@ jobs:
             exit 1
           fi
           cp "${artifacts[@]}" pypi-upload/
-
-      - name: Upload build artifacts
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        with:
-          name: pypi-${{ matrix.name }}
-          path: ${{ matrix.path }}/dist/
-          retention-days: 30
-
-      - name: Publish to PyPI
-        if: github.event_name == 'workflow_dispatch' && github.event.inputs.dry_run == 'false'
+      - name: Publish ${{ matrix.name }} to PyPI
         uses: pypa/gh-action-pypi-publish@dc37677b2e1c63e2034f94d8a5b11f265b73ba33 # v1.14.2
         with:
-          packages-dir: ${{ matrix.path }}/pypi-upload/
+          packages-dir: pypi-upload/
           skip-existing: true
 
   # -------------------------------------------------------------------
