+++ agent-governance-python/agent-compliance/pyproject.toml
@@ -31,6 +31,7 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Artificial Intelligence",
 ]
 dependencies = [
+    "agt-evidence>=5.0.0,<6.0",
     "pydantic>=2.4.0,<3.0",
     "pyyaml>=6.0,<7.0",
     "click>=8.0,<9.0",
