+++ agent-governance-python/agent-compliance/src/agent_compliance/verify.py
@@ -23,6 +23,9 @@ from typing import Any, Optional
 
 import yaml
 
+from agt_evidence import EVIDENCE_SCHEMA as EVIDENCE_SCHEMA
+from agt_evidence import RuntimeEvidence
+
 logger = logging.getLogger(__name__)
 
 # MSRC Case 112362:
@@ -33,6 +36,7 @@ ALLOWED_MODULE_PREFIXES = frozenset(
     {
         "agent_os.",
         "agentmesh.",
+        "agt_evidence.",
         "agent_compliance.",
         "agent_sre.",
         "agent_hypervisor.",
@@ -43,13 +47,15 @@ ALLOWED_MODULE_PREFIXES = frozenset(
     }
 )
 
-EVIDENCE_SCHEMA = "agt-runtime-evidence/v1"
 MAX_POLICY_FILE_BYTES = 10 * 1024 * 1024
 
 
 def _validate_module_name(mod_name: str) -> None:
     """Raise ValueError if mod_name is not in the governance module allowlist."""
-    if not any(mod_name.startswith(prefix) for prefix in ALLOWED_MODULE_PREFIXES):
+    if not any(
+        mod_name == prefix.removesuffix(".") or mod_name.startswith(prefix)
+        for prefix in ALLOWED_MODULE_PREFIXES
+    ):
         raise ValueError(
             f"Module '{mod_name}' is not in the allowed governance module list. "
             f"Only modules with prefixes {sorted(ALLOWED_MODULE_PREFIXES)} are permitted."
@@ -84,7 +90,7 @@ OWASP_ASI_CONTROLS = {
     },
     "ASI-06": {
         "name": "Insufficient Logging",
-        "module": "agentmesh.governance.audit",
+        "module": "agt_evidence.audit",
         "check": "AuditChain",
     },
     "ASI-07": {
@@ -131,49 +137,6 @@ class EvidenceCheck:
     status: str
     message: str
     observed: dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass
-class RuntimeEvidence:
-    """Runtime evidence manifest emitted by a deployment."""
-
-    source_path: str
-    schema: str
-    generated_at: str
-    toolkit_version: str
-    deployment: dict[str, Any]
-
-    @classmethod
-    def load(cls, path: str | Path) -> "RuntimeEvidence":
-        evidence_path = Path(path).expanduser().resolve()
-        raw = evidence_path.read_text(encoding="utf-8")
-
-        if evidence_path.suffix.lower() in {".yaml", ".yml"}:
-            data = yaml.safe_load(raw)
-        else:
-            data = json.loads(raw)
-
-        if not isinstance(data, dict):
-            raise ValueError("Evidence file must contain an object at the top level.")
-
-        schema = data.get("schema")
-        if schema != EVIDENCE_SCHEMA:
-            raise ValueError(
-                f"Unsupported evidence schema {schema!r}. "
-                f"Expected {EVIDENCE_SCHEMA!r}."
-            )
-
-        deployment = data.get("deployment")
-        if not isinstance(deployment, dict):
-            raise ValueError("Evidence file missing required 'deployment' object.")
-
-        return cls(
-            source_path=str(evidence_path),
-            schema=schema,
-            generated_at=str(data.get("generated_at", "")),
-            toolkit_version=str(data.get("toolkit_version", "")),
-            deployment=deployment,
-        )
 
 
 @dataclass
