+++ agent-governance-python/agent-governance-toolkit-core/pyproject.toml
@@ -39,6 +39,7 @@ classifiers = [
 # Merged dependencies from agent-os-kernel, agentmesh-primitives,
 # agentmesh-runtime, agent-hypervisor, and agentmesh-platform.
 dependencies = [
+    "agt-evidence>=5.0.0,<6.0",
     "agt-policies>=5.0.0,<6.0",
     "pydantic[email]>=2.5.0,<3.0",
     "pyyaml>=6.0,<7.0",
