+++ scripts/ci/no-custom-crypto.sh
@@ -4,6 +4,7 @@
 #
 # AGT's crypto belongs in:
 #   agent-governance-python/agent-mesh/  (identity, encryption, trust)
+#   agent-governance-python/agt-evidence/  (tamper-evident audit records)
 #   agent-governance-rust/src/crypto/
 #   agent-governance-typescript/src/encryption/
 #   agent-governance-dotnet/src/Security/
@@ -53,6 +54,7 @@ CRYPTO_PATTERNS=(
 # Paths where crypto is allowed
 ALLOWED_PATHS=(
   'agent-governance-python/agent-mesh/'
+  'agent-governance-python/agt-evidence/'
   'agent-governance-rust/src/'
   'agent-governance-typescript/src/encryption/'
   'agent-governance-dotnet/src/'
@@ -66,6 +68,7 @@ PATTERN=$(IFS='|'; echo "${CRYPTO_PATTERNS[*]}")
 ADDED=$(git diff "$BASE_REF"...HEAD --diff-filter=ACMR -U0 -- \
   '*.py' '*.ts' '*.rs' '*.cs' '*.go' \
   ':!agent-governance-python/agent-mesh/**' \
+  ':!agent-governance-python/agt-evidence/**' \
   ':!agent-governance-rust/**' \
   ':!agent-governance-typescript/src/encryption/**' \
   ':!agent-governance-dotnet/**' \
